pyproject.toml
[tool.poetry]
name = "django-blog"
version = "0.1.0"
description = "Django admin based blog"
authors = ["Charles de Freitas <charles@defreitas.io>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1"
djangorestframework = "^3.11.1"
gitpython = "^3.1.7"

[tool.poetry.dev-dependencies]
pylint = "^2.5.3"
uvicorn = "^0.11.8"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
